Updating all test cases to work with latest NTO running as daemon. Separate all tests to individual files. Running all scripts in suite.
