+++ openshift_tooling/node_tuning_operator/README.md
@@ -1,22 +1,29 @@
 # Node Tuning Operator
 
-## Test: Node Tuning Operator - Core functionality
+## Prerequisite
 
-### What is tested:
+OpenShift v 4.0 or higher 
+Python (tested with v 2.7.5) 
+
+## Test Suite
+
+To run all test at once, you can use `nto_suite` python script
+
+```bash
+python nto_suite.py
+```
+
+## Test cases
+### Test: Node Tuning Operator - Core functionality
+
+#### What is tested:
 
 - Verification that after creating new resource with 'es' label pod will be tuned
-- Verification that modification (increase) of a parameter: net.netfilter.nf_conntrack_max will take effect on every node of the cluster.
-- Verification that modification (decrease) of a parameter: net.netfilter.nf_conntrack_max will take effect on every node of the cluster.
-- Verification that modification (increase) of a parameter: kernel.pid_max will take effect on every node of the cluster.
-- Verification that modification (decrease) of a parameter: kernel.pid_max will NOT take effect on every node of the cluster.
-- Verification that after changing priority pod will be tuned.
+- Verification that modification of a parameter: kernel.pid_max will take effect on every node of the cluster.
+- Verification that removing custom tuning take effect on every node of the cluster
+- Verification number of secrets
 
-### Prerequisite
-
-OpenShift v 4.0 or higher
-Python (tested with v 2.7.5)
-
-### How to run test:
+#### How to run test:
 
 - Login to OCP
 ```bash
@@ -24,16 +31,10 @@ oc login -u <user_name> -p <password>
 ```
 - To test just run python script
 ```bash
-python node_tuning_operator.py
+python nto_test_core_functionality_is_working.py
 ```
 
-- If test pass and you don't need configuration files used during test you can delete them:
-```bash
-rm default_values_netfilter.yaml default_values_pid.yaml default_values_priority.yaml default_values.yaml new_conntract_decrease.yaml new_conntract_increase.yaml new_kernel_pid_decrease.yaml new_kernel_pid_increase.yaml new_priority.yaml
-
-```
-
-### How to recover if something goes wrong:
+#### How to recover if something goes wrong:
 Script has step to cleanup after test - even when test failed, but if something unexpected happen then please follow below steps:
 
 - Delete project:
@@ -41,16 +42,14 @@ Script has step to cleanup after test - even when test failed, but if something 
 oc delete project my-logging-project
 ```
 
-- Restore previous configuration:
+- Restore default configuration:
 ```bash
-oc delete tuned default
-oc create -f ./default_values.yaml
-
+oc delete tuned max-pid-test -n openshift-cluster-node-tuning-operator
 ```
 
-## Test: Node Tuning Operator - custom tuning is working
+### Test: Node Tuning Operator - custom tuning is working
 
-### What is tested:
+#### What is tested:
 
 - Verification if after creating custom tuning new tuning exists.
 - Verification if after creating custom tuning new tuning applied to tuned-profiles.
@@ -58,12 +57,7 @@ oc create -f ./default_values.yaml
 - Verification if correct nodes are tuned by new custom tuning.
 - Logs verification
 
-### Prerequisite
-
-OpenShift v 4.0 or higher
-Python (tested with v 2.7.5)
-
-### How to run test:
+#### How to run test:
 
 - Login to OCP
 ```bash
@@ -71,13 +65,64 @@ oc login -u <user_name> -p <password>
 ```
 - To test just run python script
 ```bash
-python node_tuning_operator_custom_tuning.py
+python nto_test_custom_tuning.py
 ```
 
-### How to recover if something goes wrong:
+#### How to recover if something goes wrong:
 Script has step to cleanup after test - even when test failed, but if something unexpected happen then please follow below steps:
 
 - Delete tuned:
 ```bash
-oc delete tuned router
+oc delete tuned router -n openshift-cluster-node-tuning-operator
+```
+
+
+### Test: Node Tuning Operator - Daemon mode - label pod
+
+#### What is tested:
+
+- Verification if labeling pod to match one of default profile will take effect on every node of cluster
+
+#### How to run test:
+
+- Login to OCP
+```bash
+oc login -u <user_name> -p <password>
+```
+- To test just run python script
+```bash
+python nto_test_daemon_mode_label_pod.py
+```
+
+#### How to recover if something goes wrong:
+Script has step to cleanup after test - even when test failed, but if something unexpected happen then please follow below steps:
+
+- Delete tuned:
+```bash
+oc label pods --all tuned.openshift.io/elasticsearch-
+```
+
+### Test: Node Tuning Operator - Daemon mode - Remove pod
+
+#### What is tested:
+
+- Verification that after deleting labeled pod, daemon will restore default values
+
+#### How to run test:
+
+- Login to OCP
+```bash
+oc login -u <user_name> -p <password>
+```
+- To test just run python script
+```bash
+python nto_test_daemon_mode_remove_pod.py
+```
+
+#### How to recover if something goes wrong:
+Script has step to cleanup after test - even when test failed, but if something unexpected happen then please follow below steps:
+
+- Delete tuned:
+```bash
+oc label pods --all tuned.openshift.io/elasticsearch-
 ```
